Exporting a function instead of class. Fixes after review.
